NEXUS-5472: WL Pool was not shut down forcedly
There might be cases, when existing code was not enough.
For example, on Bamboo a request to Codehaus prefix file
fetch took 25sec (unsure why, was codehaus misbehaving or
whatnot), but the existing code left the thread to
continue. While this did not interfere when Nexus used
as bundle (as rebooting a bundle reboots JVM), in case
of WAR deployment this might interfere.

Now adding shutdownNow call, that will interrupt
the executing threads (and should help in case like
above, where HTTP client was blocked on IO socket
op waiting for response).